Fixed the Grab bug
Make sure everything in the inspector is set up properly. But bug was due to the change in sprites since the size was now larger. Also I instead made the grabbedObject variable a GameObject not a Transform. Doesnt matter too much but I thought that was the issue and it wasn't you may change it back.